+++ custom_components/tesla_extended/const.py
@@ -1,17 +1,19 @@
 """Const file for Tesla cars."""
 
-VERSION = "1.0.0"
+VERSION = "1.2.0"
 CONF_EXPIRATION = "expiration"
 CONF_INCLUDE_VEHICLES = "include_vehicles"
 CONF_POLLING_POLICY = "polling_policy"
 CONF_WAKE_ON_START = "enable_wake_on_start"
 CONF_ENABLE_TESLAMATE = "enable_teslamate"
+CONF_SENTRY_SCAN_INTERVAL = "sentry_scan_interval"
 DOMAIN = "tesla_extended"
 ATTRIBUTION = "Data provided by Tesla"
 DATA_LISTENER = "listener"
 DEFAULT_SCAN_INTERVAL = 660
 DEFAULT_WAKE_ON_START = False
 DEFAULT_ENABLE_TESLAMATE = False
+DEFAULT_SENTRY_SCAN_INTERVAL = 660
 ERROR_URL_NOT_DETECTED = "url_not_detected"
 MIN_SCAN_INTERVAL = 10
 
@@ -26,7 +28,7 @@ PLATFORMS = [
     "switch",
     "button",
     "update",
-    "text",
+    "number",
 ]